Moving testAlertSchedulingWithDistributedDatabase as an integration test since it is now usign TSDB Service
